requirements.txt
yfinance
pandas
numpy
requests
beautifulsoup4
python-dateutil
scipy
alpaca-py>=0.30.0
websockets>=12.0
fastapi
uvicorn
sse-starlette
